requirements.txt
pynput
tqdm
pyserial
PyYAML
